requirements.txt
pygame==2.6.1
playsound==1.3.0
requests==2.32.3
obs-websocket-py==1.0
pyttsx3==2.98
streamdeck==0.9.5
pillow==10.4.0
tk==0.1.0
pyinstaller==6.10.0